+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4-alpha] - 2024-05-18
+
+### Added
+
+- Processing DAG persists arXiv research to S3 and neo4j
+- neo4j seeding script refactored to DAG
+
+### Removed
+
+- arXiv research processing Lambdas
+
+## [0.3.3-alpha] - 2024-05-16
+
+### Changed
+
+- Use single host in dev for orchestration and neo4j
+
 ## [0.3.2-alpha] - 2024-05-04
 
 ### Added